requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
flask==3.1.2
flask-cors==4.0.1
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
pillow==11.3.0
pyautogui==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
pyscreeze==1.0.1
python-dotenv==1.1.1
pytweening==1.2.0
pywhatkit==5.4
requests==2.31.0
soupsieve==2.8
typing-extensions==4.15.0
urllib3==2.5.0
werkzeug==3.1.3
wikipedia==1.4.0
zipp==3.23.0
pyinstaller==6.3.0
marshmallow==3.21.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
